chore(restructure): move hooks -> openshorts/overlays/hooks.py
Phase 1 step 3: relocate the PIL hook-card generator + FFmpeg overlay
helper. Imported by app.py and the three root verify_*.py scripts.
Adds a re-export shim at the old path. Tests stay 62/62 green.
